Revert Util changes
