+++ src/main/java/io/javalin/plugin/openapi/JavalinOpenApi.kt
@@ -3,18 +3,13 @@ package io.javalin.plugin.openapi
 import io.javalin.Javalin
 import io.javalin.core.event.HandlerMetaInfo
 import io.javalin.core.plugin.PluginNotFoundException
-import io.javalin.core.util.OptionalDependency
-import io.javalin.core.util.Util
 import io.javalin.plugin.openapi.dsl.applyMetaInfoList
 import io.javalin.plugin.openapi.dsl.updateComponents
 import io.javalin.plugin.openapi.dsl.updatePaths
 import io.javalin.plugin.openapi.jackson.JacksonModelConverterFactory
-import io.swagger.util.Yaml
 import io.swagger.v3.core.converter.ModelConverter
 import io.swagger.v3.core.converter.ModelConverters
 import io.swagger.v3.oas.models.OpenAPI
-import io.swagger.v3.parser.OpenAPIV3Parser
-import org.slf4j.LoggerFactory
 
 class CreateSchemaOptions(
         val handlerMetaInfoList: List<HandlerMetaInfo>,
@@ -33,8 +28,6 @@ class CreateSchemaOptions(
 )
 
 object JavalinOpenApi {
-    private val logger = LoggerFactory.getLogger(JavalinOpenApi::javaClass.get())
-
     @JvmStatic
     fun createSchema(javalin: Javalin): OpenAPI {
         val handler = try {
@@ -49,7 +42,7 @@ object JavalinOpenApi {
     fun createSchema(options: CreateSchemaOptions): OpenAPI {
         val baseConfiguration = options.initialConfigurationCreator.create()
         val modelConverter = options.modelConverterFactory.create()
-        val schema = runWithModelConverter(modelConverter) {
+        return runWithModelConverter(modelConverter) {
             baseConfiguration.apply {
                 updateComponents {
                     applyMetaInfoList(options.handlerMetaInfoList, options)
@@ -59,20 +52,6 @@ object JavalinOpenApi {
                 }
             }
         }
-
-        Util.executeIfDependencyIsPresent(OptionalDependency.SWAGGERPARSER) {
-            val parsedSchema = OpenAPIV3Parser().readContents(Yaml.mapper().writeValueAsString(schema))
-
-            if(parsedSchema.messages.isNotEmpty()){
-                logger.warn("The generated OpenApi specification is not valid")
-
-                parsedSchema.messages.forEach {
-                    logger.warn(it)
-                }
-            }
-        }
-
-        return schema
     }
 }
