+++ src/main/java/io/javalin/plugin/openapi/OpenApiHandler.kt
@@ -3,11 +3,18 @@ package io.javalin.plugin.openapi
 import io.javalin.Javalin
 import io.javalin.core.event.HandlerMetaInfo
 import io.javalin.core.util.Header
+import io.javalin.core.util.OptionalDependency
+import io.javalin.core.util.Util
 import io.javalin.http.Context
 import io.javalin.http.Handler
 import io.javalin.plugin.openapi.annotations.ContentType
 import io.javalin.plugin.openapi.annotations.OpenApi
+import io.swagger.util.Yaml
 import io.swagger.v3.oas.models.OpenAPI
+import io.swagger.v3.parser.OpenAPIV3Parser
+import org.slf4j.LoggerFactory
+
+private val logger = LoggerFactory.getLogger(OpenApiHandler::javaClass.javaClass)
 
 class OpenApiHandler(app: Javalin, val options: OpenApiOptions) : Handler {
     private val handlerMetaInfoList = mutableListOf<HandlerMetaInfo>()
@@ -22,15 +29,32 @@ class OpenApiHandler(app: Javalin, val options: OpenApiOptions) : Handler {
         }
     }
 
-    fun createOpenAPISchema(): OpenAPI = JavalinOpenApi.createSchema(
-        CreateSchemaOptions(
-            handlerMetaInfoList = handlerMetaInfoList,
-            initialConfigurationCreator = options.initialConfigurationCreator,
-            default = options.default,
-            modelConverterFactory = options.modelConverterFactory,
-            packagePrefixesToScan = options.packagePrefixesToScan
+    fun createOpenAPISchema(): OpenAPI {
+        val schema = JavalinOpenApi.createSchema(
+            CreateSchemaOptions(
+                handlerMetaInfoList = handlerMetaInfoList,
+                initialConfigurationCreator = options.initialConfigurationCreator,
+                default = options.default,
+                modelConverterFactory = options.modelConverterFactory,
+                packagePrefixesToScan = options.packagePrefixesToScan
+            )
         )
-    )
+
+        if(options.validateSchema){
+            Util.ensureDependencyPresent(OptionalDependency.SWAGGERPARSER)
+            val parsedSchema = OpenAPIV3Parser().readContents(Yaml.mapper().writeValueAsString(schema))
+
+            if(parsedSchema.messages.isNotEmpty()){
+                logger.warn("The generated OpenApi specification is not valid")
+
+                parsedSchema.messages.forEach {
+                    logger.warn(it)
+                }
+            }
+        }
+
+        return schema
+    }
 
     // This function is synchronized because an attacker can trigger the openapi schema generation very often
     // It is ensured, that the schema is generated only maximal once after adding handlers
